Extract functions from analyze.js, to make them accessible to tests
